Modify index.html and style.css
Change class 'main-title' to an existing 'title' and add modifiers 'title--black', 'title--margin-bottom-8px' and 'title--margin-bottom-12px'.

Therefore, for block 'intro' class 'intro__title' describes the typography of the header and positioning inherent in this unit content.

Modifiers used for additional styles, which are different headings on the page, not to combine a variety of classes.

Also changed 'intro__post' and 'education__post' on modifiers 'post--margin-bottom-20px' and 'post--margin-bottom-12px', respectively.

In style.css styles are added for 'intro' block and styles for the previously specified modifiers.