Update index.html and style.css
Made styling block 'Projects'.

In index.html added the span element
class 'projects__item-inner', which wraps the entire content of the list item.
This is done in order to be able to separately style the numeric bullets from surrounding content.

!!! ATTENTION !!!
Used pseudo-class '::marker'.
He is'nt support in Internet Explorer, Samsung Internet; Partial support in Safari, Safari on iOS.

Also added the class 'projects__brackets' for the design of the styles '[' and ']'.